requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.2.0
PyJWT==1.7.1
python-memcached==1.59
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
